backend/requirements.txt
fastapi
pydantic
pymongo
uvicorn
python-jose
passlib
python-multipart
email-validator
pytest
pytest-asyncio
httpx
dotenv
motorfastapi
httpx
pandas
numpy
faiss-cpu
sentence-transformers
uvicorn
